Renamed app to SEA-PHAGES
Renamed the app and made several minor changes to npm package versions.
Builds successfully in development and production for Android.
