+++ SEAPHAGES_App_RN/.babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": ["react-native"]
+  "presets": ["react-native"],
+  "plugins": ["transform-decorators-legacy"]
 }
