+++ SEAPHAGES_App_RN/android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.seaphages_app_rn",
+    package = "org.cresawnlab.seaphages_app_rn",
 )
 
 android_resource(
     name = "res",
-    package = "com.seaphages_app_rn",
+    package = "org.cresawnlab.seaphages_app_rn",
     res = "src/main/res",
 )
