+++ SEAPHAGES_App_RN/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">SEAPHAGES_App_RN</string>
+    <string name="app_name">SEA-PHAGES</string>
 </resources>
