+++ SEAPHAGES_App_RN/android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'SEAPHAGES_App_RN'
+rootProject.name = 'SEAPHAGES'
 include ':react-native-splash-screen'
 project(':react-native-splash-screen').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-splash-screen/android')
 include ':react-native-maps'
