+++ SEAPHAGES_App_RN/app.json
@@ -1,4 +1,4 @@
 {
   "name": "SEAPHAGES_App_RN",
-  "displayName": "SEAPHAGES_App_RN"
+  "displayName": "SEA-PHAGES"
 }
