+++ SEAPHAGES_App_RN/app/config/settings.js
@@ -3,7 +3,7 @@
 let METEOR_URL = 'ws://phamerator.org:3000/websocket';
 //let METEOR_URL = 'ws://localhost:3000/websocket';
 if (process.env.NODE_ENV === 'production') {
-    METEOR_URL = ''; // your production server url
+    METEOR_URL = 'ws://phamerator.org:3000/websocket'; // your production server url
 }
 
 export const settings = {
