+++ SEAPHAGES_App_RN/app/index.js
@@ -2,7 +2,7 @@
 
 import React, { Component } from 'react';
 import Meteor from 'react-native-meteor';
-import SplashScreen from 'react-native-splash-screen';
+//import SplashScreen from 'react-native-splash-screen';
 
 import { Root } from './config/routes';
 import settings from './config/settings';
@@ -11,7 +11,7 @@ Meteor.connect(settings.METEOR_URL);
 
 class SEAPHAGES_App_RN extends Component {
     componentDidMount() {
-        SplashScreen.hide();
+        //SplashScreen.hide();
     }
 
     render() {
