+++ SEAPHAGES_App_RN/ios/SEAPHAGES_App_RN/AppDelegate.m
@@ -8,7 +8,6 @@
  */
 
 #import "AppDelegate.h"
-#import "SplashScreen.h"
 
 #import <React/RCTBundleURLProvider.h>
 #import <React/RCTRootView.h>
@@ -32,9 +31,6 @@
   rootViewController.view = rootView;
   self.window.rootViewController = rootViewController;
   [self.window makeKeyAndVisible];
-  
-  [SplashScreen show];
-  
   return YES;
 }
