+++ SEAPHAGES_App_RN/package.json
@@ -8,18 +8,19 @@
   },
   "dependencies": {
     "moment": "^2.18.1",
-    "native-base": "^2.1.5",
-    "react": "16.0.0-alpha.12",
+    "native-base": "^2.2.1",
+    "react": "^16.0.0-alpha.12",
     "react-native": "^0.46.2",
     "react-native-easy-grid": "^0.1.14",
-    "react-native-maps": "^0.15.2",
+    "react-native-maps": "^0.15.3",
     "react-native-meteor": "^1.0.3",
-    "react-native-splash-screen": "^3.0.0",
+    "react-native-splash-screen": "^2.1.0",
     "react-navigation": "^1.0.0-beta.11"
   },
   "devDependencies": {
     "babel-jest": "20.0.3",
-    "babel-preset-react-native": "1.9.2",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-preset-react-native": "^1.9.2",
     "jest": "20.0.4",
     "react-native-cli": "^2.0.1",
     "react-test-renderer": "16.0.0-alpha.12"
